little restructuring
